pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quasar-mdiqkd"
version = "0.1.0"
description = "Production-grade MDI-QKD simulation stack built on SeQUeNCe extensions"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Naman Srivastava" }]
license = { text = "Apache-2.0" }
keywords = ["quantum", "QKD", "MDI-QKD", "simulation", "SeQUeNCe", "networking"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering",
]

dependencies = [
  "pydantic>=2.8",
  "pyyaml>=6.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=16.1",
  "matplotlib>=3.8",
  "networkx>=3.3",
  "scipy>=1.13",
  "tqdm>=4.66",
  "streamlit>=1.36",
  "plotly>=5.23",
  "shapely>=2.0",
  "rich>=13.7",
  "typer>=0.12",
  "loguru>=0.7",
  # geopandas is optional on Windows; users can install via extra below
]

[project.optional-dependencies]
geo = ["geopandas>=1.0"]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5.0",
  "mypy>=1.10",
  "types-PyYAML>=6.0.12",
  "types-setuptools",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sequence_ext*", "configs*"]

# If you later decide to expose CLI entry points (e.g., `quasar-run`),
# add __init__.py to the `scripts/` folder to make it a package and then
# uncomment the following block, pointing to the scripts' main() functions.
# [project.scripts]
# quasar-run = "scripts.run_scenario:main"
# quasar-sweep = "scripts.sweep:main"
# quasar-report = "scripts.report:main"

# ----------------------------
# Tooling configuration
# ----------------------------

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --disable-warnings"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["sequence_ext"]
omit = ["**/__init__.py"]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
]
ignore = [
  "E203",  # whitespace before ':', conflicts with black style if used later
]

[tool.ruff.isort]
known-first-party = ["sequence_ext", "configs"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
strict_optional = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Streamlit-specific note:
# If Streamlit is not installed in CI, you can skip dashboard tests or
# install the base 'project.dependencies' which includes streamlit.
